+++ native/packages/react-native-bpk-component-button/src/layout/BpkButtonContainer.ios.js
@@ -17,8 +17,8 @@
  */
 
 import {
-  View,
   TouchableHighlight,
+  ViewPropTypes,
 } from 'react-native';
 import React from 'react';
 import PropTypes from 'prop-types';
@@ -82,7 +82,7 @@ BpkButtonContainer.propTypes = {
   disabled: PropTypes.bool,
   iconOnly: PropTypes.bool,
   large: PropTypes.bool,
-  style: View.propTypes.style,
+  style: ViewPropTypes.style,
   theme: themePropType,
 };